// server/src/domain/enums/routes.enum.ts
export enum SeekerRoutes {
  PROFILE = '/profile',
  PROFILE_EXPERIENCES = '/profile/experiences',
  PROFILE_EXPERIENCES_ID = '/profile/experiences/:experienceId',
  PROFILE_EDUCATION = '/profile/education',
  PROFILE_EDUCATION_ID = '/profile/education/:educationId',
  PROFILE_SKILLS = '/profile/skills',
  PROFILE_LANGUAGES = '/profile/languages',
  PROFILE_RESUME = '/profile/resume',
  PROFILE_AVATAR = '/profile/avatar',
  PROFILE_BANNER = '/profile/banner',
  APPLICATIONS = '/applications',
  APPLICATIONS_ANALYZE_RESUME = '/applications/analyze-resume',
  APPLICATIONS_ID = '/applications/:id',
}